Consolidate jl.config.js injection logic into jl.install.js
